# apps/zero-api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/drafts": {
      "post": {
        "operationId": "DraftsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDraftDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftDto"
                }
              }
            }
          }
        },
        "tags": [
          "drafts"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/drafts/{id}": {
      "get": {
        "operationId": "DraftsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftDto"
                }
              }
            }
          }
        },
        "tags": [
          "drafts"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "put": {
        "operationId": "DraftsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDraftDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftDto"
                }
              }
            }
          }
        },
        "tags": [
          "drafts"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "DraftsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftDto"
                }
              }
            }
          }
        },
        "tags": [
          "drafts"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Energy Web Zero API",
    "description": "",
    "version": "0.3",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3333"
    },
    {
      "url": "http://zero-dev.energyweb.org"
    },
    {
      "url": "https://zero-dev.energyweb.org"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "DraftType": {
        "type": "string",
        "enum": [
          "facility"
        ]
      },
      "DraftDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "userId": {
            "type": "number",
            "example": 1
          },
          "data": {
            "type": "object",
            "example": "{ \"field1\": \"value\", \"field2\" : [1, 2, 3] }"
          },
          "draftType": {
            "$ref": "#/components/schemas/DraftType"
          }
        },
        "required": [
          "id",
          "userId",
          "data",
          "draftType"
        ]
      },
...
      "CreateDraftDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object"
          },
          "draftType": {
            "$ref": "#/components/schemas/DraftType"
          }
        },
        "required": [
          "data",
          "draftType"
        ]
      },
      "UpdateDraftDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object"
          }
        }
      },
...
    }
  }
}